# api-spec.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/v1/profiles/talent":  {
                                                  "post":  {
                                                               "operationId":  "ProfilesController_createTalentProfile",
                                                               "parameters":  [

                                                                              ],
                                                               "requestBody":  {
                                                                                   "required":  true,
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/CreateTalentProfileDto"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               },
                                                               "responses":  {
                                                                                 "201":  {
                                                                                             "description":  ""
                                                                                         }
                                                                             }
                                                           }
                                              },
                  "/api/v1/profiles/agency":  {
                                                  "post":  {
                                                               "operationId":  "ProfilesController_createAgencyProfile",
                                                               "parameters":  [

                                                                              ],
                                                               "requestBody":  {
                                                                                   "required":  true,
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/CreateAgencyProfileDto"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               },
                                                               "responses":  {
                                                                                 "201":  {
                                                                                             "description":  ""
                                                                                         }
                                                                             }
                                                           }
                                              },
                  "/api/v1/profiles/talent/{id}":  {
                                                       "get":  {
                                                                   "operationId":  "ProfilesController_getTalentProfile",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "required":  true,
                                                                                          "in":  "path",
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  ""
                                                                                             }
                                                                                 }
                                                               },
                                                       "put":  {
                                                                   "operationId":  "ProfilesController_updateTalentProfile",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "required":  true,
                                                                                          "in":  "path",
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "required":  true,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/UpdateTalentProfileDto"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  ""
                                                                                             }
                                                                                 }
                                                               }
                                                   },
                  "/api/v1/profiles/public/talent/{id}":  {
                                                              "get":  {
                                                                          "operationId":  "ProfilesController_getPublicTalentProfile",
                                                                          "parameters":  [
                                                                                             {
                                                                                                 "name":  "id",
                                                                                                 "required":  true,
                                                                                                 "in":  "path",
                                                                                                 "schema":  {
                                                                                                                "type":  "string"
                                                                                                            }
                                                                                             }
                                                                                         ],
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  ""
                                                                                                    }
                                                                                        }
                                                                      }
                                                          },
                  "/api/v1/profiles/public/talents":  {
                                                          "get":  {
                                                                      "operationId":  "ProfilesController_getPublicTalents",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "page",
                                                                                             "required":  true,
                                                                                             "in":  "query",
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         },
                                                                                         {
                                                                                             "name":  "limit",
                                                                                             "required":  true,
                                                                                             "in":  "query",
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         },
                                                                                         {
                                                                                             "name":  "skills",
                                                                                             "required":  true,
                                                                                             "in":  "query",
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  ""
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
                  "/api/v1/profiles/talents":  {
                                                   "get":  {
                                                               "operationId":  "ProfilesController_getTalents",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "page",
                                                                                      "required":  true,
                                                                                      "in":  "query",
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "limit",
                                                                                      "required":  true,
                                                                                      "in":  "query",
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "skills",
                                                                                      "required":  true,
                                                                                      "in":  "query",
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  ""
                                                                                         }
                                                                             }
                                                           }
                                               },
                  "/api/v1/profiles/agency/{id}":  {
                                                       "get":  {
                                                                   "operationId":  "ProfilesController_getAgencyProfile",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "required":  true,
                                                                                          "in":  "path",
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  ""
                                                                                             }
                                                                                 }
                                                               },
                                                       "put":  {
                                                                   "operationId":  "ProfilesController_updateAgencyProfile",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "required":  true,
                                                                                          "in":  "path",
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "required":  true,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/UpdateAgencyProfileDto"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  ""
                                                                                             }
                                                                                 }
                                                               }
                                                   },
                  "/api/v1/profiles/talent/me":  {
                                                     "put":  {
                                                                 "operationId":  "ProfilesController_updateMyTalentProfile",
                                                                 "parameters":  [

                                                                                ],
                                                                 "requestBody":  {
                                                                                     "required":  true,
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/UpdateTalentProfileDto"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 },
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  ""
                                                                                           }
                                                                               }
                                                             }
                                                 },
                  "/api/v1/profiles/{id}":  {
                                                "delete":  {
                                                               "operationId":  "ProfilesController_deleteProfile",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "required":  true,
                                                                                      "in":  "path",
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "204":  {
                                                                                             "description":  ""
                                                                                         }
                                                                             }
                                                           }
                                            },
...
              },
    "info":  {
                 "title":  "Castlyo API",
                 "description":  "Professional casting and talent matching platform API",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
                       "schemas":  {
...
                                       "CreateTalentProfileDto":  {
                                                                      "type":  "object",
                                                                      "properties":  {

                                                                                     }
                                                                  },
                                       "CreateAgencyProfileDto":  {
                                                                      "type":  "object",
                                                                      "properties":  {

                                                                                     }
                                                                  },
                                       "UpdateTalentProfileDto":  {
                                                                      "type":  "object",
                                                                      "properties":  {

                                                                                     }
                                                                  },
                                       "UpdateAgencyProfileDto":  {
                                                                      "type":  "object",
                                                                      "properties":  {

                                                                                     }
                                                                  },
...
                                   }
                   }
}